moved network extension methods to extensions folder moved iterators to iterators folder
