+++ TLM/TLM/Patch/_RoadBaseAI/UpdateLanesPatch.cs
@@ -3,6 +3,7 @@ namespace TrafficManager.Patch._RoadBaseAI {
     using JetBrains.Annotations;
     using TrafficManager.State;
     using TrafficManager.Util;
+    using TrafficManager.Util.Extensions;
 
     [HarmonyPatch(typeof(RoadBaseAI), "UpdateLanes")]
     [UsedImplicitly]
