+++ TLM/TLM/Patch/_VehicleAI/_CarAI/CalculateSegmentPositionPatch2.cs
@@ -8,6 +8,7 @@ namespace TrafficManager.Patch._VehicleAI._CarAI {
     using Manager.Impl;
     using State;
     using State.ConfigData;
+    using TrafficManager.Util.Extensions;
     using UnityEngine;
     using Util;
