+++ TLM/TLM/TLM.csproj
@@ -146,10 +146,10 @@
     <Compile Include="Manager\Impl\ExtSegmentEndManager.cs" />
     <Compile Include="Manager\Impl\ExtSegmentManager.cs" />
     <Compile Include="Manager\Impl\GeometryNotifier.cs" />
-    <Compile Include="Manager\Impl\GetNodeSegmentIdsEnumerable.cs" />
-    <Compile Include="Manager\Impl\GetNodeSegmentIdsEnumerator.cs" />
-    <Compile Include="Manager\Impl\GetSegmentLaneIdsEnumerable.cs" />
-    <Compile Include="Manager\Impl\GetSegmentLaneIdsEnumerator.cs" />
+    <Compile Include="Util\Iterators\GetNodeSegmentIdsEnumerable.cs" />
+    <Compile Include="Util\Iterators\GetNodeSegmentIdsEnumerator.cs" />
+    <Compile Include="Util\Iterators\GetSegmentLaneIdsEnumerable.cs" />
+    <Compile Include="Util\Iterators\GetSegmentLaneIdsEnumerator.cs" />
     <Compile Include="Manager\Impl\LaneIdAndIndex.cs" />
     <Compile Include="Manager\Impl\LanePos.cs" />
     <Compile Include="Manager\Impl\SpeedLimitManagerExt.cs" />
@@ -424,6 +424,7 @@
     <Compile Include="Util\Extensions\VehicleExtensions.cs" />
     <Compile Include="Util\InGameUtil.cs" />
     <Compile Include="Util\IntVector2.cs" />
+    <Compile Include="Util\LaneUtil.cs" />
     <Compile Include="Util\RoadFamilyUtil.cs" />
     <Compile Include="Util\Spiral.cs" />
     <Compile Include="Util\Caching\CameraTransformValue.cs" />
