+++ TLM/TLM/UI/SubTools/SpeedLimits/Overlay/OverlaySegmentSpeedlimitHandle.cs
@@ -1,4 +1,4 @@
-﻿namespace TrafficManager.UI.SubTools.SpeedLimits.Overlay {
+namespace TrafficManager.UI.SubTools.SpeedLimits.Overlay {
     using System;
     using ColossalFramework;
     using CSUtil.Commons;
@@ -7,6 +7,7 @@
     using TrafficManager.Manager.Impl;
     using TrafficManager.State;
     using TrafficManager.Util;
+    using TrafficManager.Util.Extensions;
 
     /// <summary>
     /// Describes a recently rendered speed icon on the speed limits overlay for SEGMENT.
