+++ TLM/TLM/UI/TrafficManagerTool.cs
@@ -31,6 +31,7 @@ namespace TrafficManager.UI {
     using static TrafficManager.Util.SegmentTraverser;
     using TrafficManager.UI.Textures;
     using TrafficManager.State.Keybinds;
+    using TrafficManager.Util.Extensions;
 
     [UsedImplicitly]
     public class TrafficManagerTool
