+++ TLM/TLM/Util/Extensions/BuildingExtensions.cs
@@ -1,5 +1,10 @@
 namespace TrafficManager.Util.Extensions {
+    using ColossalFramework;
     public static class BuildingExtensions {
+        private static Building[] _buildingBuffer = Singleton<BuildingManager>.instance.m_buildings.m_buffer;
+
+        internal static ref Building ToBuilding(this ushort buildingId) => ref _buildingBuffer[buildingId];
+
         /// <summary>
         /// Checks if the building is Created, but neither Collapsed nor Deleted.
         /// </summary>
