+++ TLM/TLM/Util/Extensions/CitizenInstanceExtensions.cs
@@ -1,5 +1,9 @@
 namespace TrafficManager.Util.Extensions {
+    using ColossalFramework;
     public static class CitizenInstanceExtensions {
+        private static CitizenInstance[] _citizenInstanceBuffer = Singleton<CitizenManager>.instance.m_instances.m_buffer;
+
+        internal static ref CitizenInstance ToCitizenInstance(this ushort citizenInstance) => ref _citizenInstanceBuffer[citizenInstance];
         /// <summary>
         /// Checks if the citizenInstance is Created, but not Deleted.
         /// </summary>
