+++ TLM/TLM/Util/Iterators/GetSegmentLaneIdsEnumerable.cs
@@ -1,7 +1,8 @@
-using System.Collections;
-using System.Collections.Generic;
+namespace TrafficManager.Util.Iterators {
+    using System.Collections;
+    using System.Collections.Generic;
+    using TrafficManager.Manager.Impl;
 
-namespace TrafficManager.Manager.Impl {
     /// <summary>
     /// Struct enumerable to be returned from the GetSegmentLaneIds method.
     /// This implementation is just for perf optimizations and should be handled with care since it returns a mutable struct!
