+++ TLM/TLM/Util/Record/LaneArrowsRecord.cs
@@ -5,6 +5,7 @@ namespace TrafficManager.Util.Record {
     using TrafficManager.API.Traffic.Enums;
     using TrafficManager.Manager.Impl;
     using TrafficManager.State;
+    using TrafficManager.Util.Extensions;
     using static TrafficManager.Util.Shortcuts;
 
     [Serializable]
