+++ TLM/TLM/Util/Record/LaneConnectionRecord.cs
@@ -5,6 +5,7 @@ namespace TrafficManager.Util.Record {
     using System.Linq;
     using TrafficManager.Manager.Impl;
     using TrafficManager.State;
+    using TrafficManager.Util.Extensions;
     using static TrafficManager.Util.Shortcuts;
 
     [Serializable]
