+++ TLM/TLM/Util/Record/VehicleRestrictionsLaneRecord.cs
@@ -4,6 +4,7 @@ namespace TrafficManager.Util.Record {
     using TrafficManager.API.Traffic.Enums;
     using TrafficManager.Manager.Impl;
     using TrafficManager.State;
+    using TrafficManager.Util.Extensions;
     using static TrafficManager.Util.Shortcuts;
 
     [Serializable]
